# api/openapi.yaml
openapi: 3.0.3
info:
  title: Assay API
  description: LLM 渠道检测平台 API。本文件是前后端契约的唯一真相源（契约优先）。
  version: 0.15.0
servers:
  - url: /api
paths:
...
  /channels/{id}:
    get:
      operationId: getChannel
      summary: 渠道详情（需 channels 权限；含模型条目列表）
      tags: [channels]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: 渠道详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateChannel
      summary: 修改渠道基本信息（需 channels 权限；apiKey 缺省=不改、给值=覆盖，写后不可回读）
      tags: [channels]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: 修改后的渠道
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 渠道名已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteChannel
      summary: 删除渠道（需 channels 权限；硬删，历史任务持有参数快照不受影响，key 不可恢复）
      tags: [channels]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: 删除成功（连同其模型条目）
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /stability/probes:
    get:
      operationId: listStabilityProbes
      summary: 稳定性检测项注册表（需 stability 权限；产出性能指标，无评分检查点）
      tags: [stability]
      responses:
        "200":
          description: 全部已注册稳定性检测项
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StabilityProbeInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /stability/tasks:
    post:
      operationId: createStabilityTask
      summary: 创建稳定性检测任务（需 stability 权限；创建时快照检测对象+协议，入队后由 worker 发压）
      tags: [stability]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StabilityTaskCreate"
      responses:
        "201":
          description: 任务已创建并入队
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StabilityTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 渠道已停用或不支持所选协议
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listStabilityTasks
      summary: 稳定性检测任务历史（需 stability 权限；按创建时间倒序，可按状态/渠道筛选）
      tags: [stability]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          description: 按任务状态筛选；缺省不过滤
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: channelId
          in: query
          description: 按渠道筛选；缺省不过滤（渠道删除后其任务 channel_id 置空，仅在不过滤时可见）
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 任务列表分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StabilityTaskList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /stability/tasks/{id}:
    get:
      operationId: getStabilityTask
      summary: 稳定性任务详情（需 stability 权限；含参数快照）
      tags: [stability]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StabilityTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /stability/tasks/{id}/export:
    get:
      operationId: exportStabilityTask
      summary: 导出稳定性报告（需 stability 权限；json=证据链自足报告，含全量逐请求样本；仅终态任务可用）
      tags: [stability]
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json]
      responses:
        "200":
          description: 报告文件（Content-Disposition attachment）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StabilityExport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 任务尚未结束，暂无报告
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /stability/tasks/{id}/cancel:
    post:
      operationId: cancelStabilityTask
      summary: 取消稳定性任务（需 stability 权限；排队或运行中的任务可取消，运行中的中止在途请求）
      tags: [stability]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: 取消后的任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StabilityTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 任务已结束，不可取消
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /stability/tasks/{id}/events:
    get:
      operationId: streamStabilityTaskEvents
      summary: 稳定性任务进度事件流（需 stability 权限；SSE，连接后先推当前快照，任务终态后关闭）
      tags: [stability]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: text/event-stream，每帧 data 为 TaskProgressEvent 的 JSON
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未登录或会话已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 无权访问该模块
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Currency:
      description: 渠道定价币种（仅标注，不做汇率换算）
      type: string
      enum: [USD, CNY]
    Channel:
      type: object
      required:
        [id, name, baseUrl, keyPrefix, protocols, currency, note, disabled, modelCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        baseUrl:
          description: 填到版本段（如 https://api.example.com/v1），平台按协议拼固定末段
          type: string
        keyPrefix:
          description: API key 脱敏前缀（key 写后不可回读）
          type: string
        protocols:
          type: array
          items:
            $ref: "#/components/schemas/Protocol"
        currency:
          $ref: "#/components/schemas/Currency"
        note:
          type: string
        disabled:
          description: 停用后不可选入新检测任务，配置与 key 保留
          type: boolean
        modelCount:
          type: integer
        lastTest:
          $ref: "#/components/schemas/ConnectivityTest"
        probeIntervalMinutes:
          description: 定时探活间隔分钟（1-1440）；缺省=未开启
          type: integer
          minimum: 1
          maximum: 1440
        probeModelId:
          description: 定时探活用的模型条目；条目被删自动置空、调度停摆
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ChannelDetail:
      allOf:
        - $ref: "#/components/schemas/Channel"
        - type: object
          required: [models]
          properties:
            models:
              type: array
              items:
                $ref: "#/components/schemas/ModelEntry"
...
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        baseUrl:
          type: string
          minLength: 1
          maxLength: 512
        apiKey:
          description: 缺省=不改；给值=整体覆盖
          type: string
          minLength: 1
        protocols:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Protocol"
        currency:
          $ref: "#/components/schemas/Currency"
        note:
          type: string
          maxLength: 500
        disabled:
          type: boolean
        probeIntervalMinutes:
          description: 定时探活间隔分钟；须与 probeModelId 成对提交整体覆盖，0=关闭（coalesce 语义写不了 null，故用 0 哨兵）
          type: integer
          minimum: 0
          maximum: 1440
        probeModelId:
          description: 定时探活模型条目（须属本渠道）；probeIntervalMinutes=0 时可缺省
          type: string
          format: uuid
    ModelEntry:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        inputPrice:
          description: 每 1M token 输入单价（币种见渠道）；三个价格字段全缺省=未定价
          type: number
        outputPrice:
          type: number
        cachedInputPrice:
          type: number
        createdAt:
          type: string
          format: date-time
...
    TaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    StabilityProbeInfo:
      description: 稳定性检测项元数据（产出性能指标，无评分检查点）
      type: object
      required: [id, name, description, protocols, estRequests]
      properties:
        id:
          description: 检测项唯一标识（如 concurrency_ladder）
          type: string
        name:
          type: string
        description:
          type: string
        protocols:
          description: 适用协议（空=全适用，前端展示为三协议皆可）
          type: array
          items:
            $ref: "#/components/schemas/Protocol"
        estRequests:
          description: 按默认参数的最坏预估请求数（前端据实选参数自行重算展示）
          type: integer
...
    StabilityTaskCreate:
      type: object
      required: [channelId, modelEntryId, probes, params]
      properties:
        channelId:
          type: string
          format: uuid
        modelEntryId:
          type: string
          format: uuid
        probes:
          description: 勾选的稳定性检测项 id 集合
          type: array
          minItems: 1
          items:
            type: string
        params:
          $ref: "#/components/schemas/StabilityTaskParams"
    StabilityTask:
      type: object
      required: [id, status, target, probes, params, progressTotal, progressDone, createdAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/TaskStatus"
        target:
          $ref: "#/components/schemas/TaskTarget"
        probes:
          type: array
          items:
            type: string
        params:
          description: 生效参数（缺省项已填入默认值）
          $ref: "#/components/schemas/StabilityTaskParams"
        progressTotal:
          type: integer
        progressDone:
          type: integer
        error:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    StabilityTaskList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StabilityTask"
        total:
          type: integer
...
    StabilityExport:
      description: 证据链自足的 JSON 导出：任务快照 + 指标报告 + 全量逐请求样本
      type: object
      required: [tool, version, task, report, samples]
      properties:
        tool:
          type: string
        version:
          type: string
        task:
          $ref: "#/components/schemas/StabilityTask"
        report:
          $ref: "#/components/schemas/StabilityReport"
        samples:
          type: array
          items:
            $ref: "#/components/schemas/StabilitySample"